+++ src/test/java/be/ugent/idlab/knows/dataio/iterator/CSVIteratorTest.java
@@ -8,6 +8,7 @@ import be.ugent.idlab.knows.dataio.record.CSVRecord;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
+import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -18,7 +19,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 public class CSVIteratorTest extends TestCore {
     @Test
     public void evaluate_0000_CSV() throws Exception {
-        Access access = makeLocalAccess("/csv/0000.csv", "", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/0000.csv", "", "csv", StandardCharsets.UTF_8);
         try (CSVSourceIterator iterator = new CSVSourceIterator(access)) {
             assertTrue(evaluate_0000(iterator));
         }
@@ -26,7 +27,7 @@ public class CSVIteratorTest extends TestCore {
 
     @Test
     public void evaluate_0001_CSV() throws Exception {
-        Access access = makeLocalAccess("/csv/0001.csv", "", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/0001.csv", "", "csv", StandardCharsets.UTF_8);
         try (CSVSourceIterator iterator = new CSVSourceIterator(access)) {
             assertTrue(evaluate_0001(iterator));
         }
@@ -35,7 +36,7 @@ public class CSVIteratorTest extends TestCore {
     @Test
     @Disabled
     public void evaluate_1001_header_col_missing_CSV() throws Exception {
-        Access access = makeLocalAccess("/csv/1001_header_col_missing.csv", "", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/1001_header_col_missing.csv", "", "csv", StandardCharsets.UTF_8);
         try (CSVSourceIterator csvSourceIterator = new CSVSourceIterator(access)) {
             //TODO should fail, check if it does
         }
@@ -43,7 +44,7 @@ public class CSVIteratorTest extends TestCore {
 
     @Test
     public void evaluate_1001_header_long_CSV() throws Exception {
-        Access access = makeLocalAccess("/csv/1001_header_long.csv", "", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/1001_header_long.csv", "", "csv", StandardCharsets.UTF_8);
         try (CSVSourceIterator iterator = new CSVSourceIterator(access)) {
             assertTrue(evaluate_1001_header_long(iterator));
         }
@@ -51,7 +52,7 @@ public class CSVIteratorTest extends TestCore {
 
     @Test
     public void evaluate_1001_header_short_CSV() throws Exception {
-        Access access = makeLocalAccess("/csv/1001_header_short.csv", "", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/1001_header_short.csv", "", "csv", StandardCharsets.UTF_8);
         try (CSVSourceIterator iterator = new CSVSourceIterator(access)) {
             assertTrue(evaluate_1001_header_short(iterator));
         }
@@ -59,7 +60,7 @@ public class CSVIteratorTest extends TestCore {
 
     @Test
     public void evaluate_0002_BOM_CSV() throws Exception {
-        Access access = makeLocalAccess("/csv/0002_BOM.csv", "", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/0002_BOM.csv", "", "csv", StandardCharsets.UTF_8);
         try (CSVSourceIterator iterator = new CSVSourceIterator(access)) {
             assertTrue(evaluate_0002_BOM(iterator));
         }
@@ -67,7 +68,7 @@ public class CSVIteratorTest extends TestCore {
 
     @Test
     public void evaluate_empty_CSV() throws Exception {
-        Access access = makeLocalAccess("/csv/empty.csv", "", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/empty.csv", "", "csv", StandardCharsets.UTF_8);
         try (CSVSourceIterator iterator = new CSVSourceIterator(access)) {
             assertTrue(evaluate_empty(iterator));
         }
@@ -75,7 +76,7 @@ public class CSVIteratorTest extends TestCore {
 
     @Test
     public void evaluateSparseInput() throws Exception {
-        Access access = makeLocalAccess("/csv/sparseInput.csv", "", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/sparseInput.csv", "", "csv", StandardCharsets.UTF_8);
         try (CSVSourceIterator iterator = new CSVSourceIterator(access)) {
             assertTrue(iterator.hasNext());
 
@@ -122,6 +123,19 @@ public class CSVIteratorTest extends TestCore {
             CSVRecord record = (CSVRecord) iterator.next();
 
             assertEquals("BO", record.get("\"ISO 3166\"").get(0));
+        }
+    }
+
+    @Test
+    public void test_big_csv() throws Exception {
+        Access access = new LocalFileAccess("csv/big.csv", "src/test/resources", "csv");
+        try (CSVSourceIterator iterator = new CSVSourceIterator(access)) {
+            int counter = 0;
+            while (iterator.hasNext()) {
+                iterator.next();
+                counter++;
+            }
+            assertEquals(58540, counter);
         }
     }
 }