+++ src/test/java/be/ugent/idlab/knows/dataio/stream/CSVStreamTest.java
@@ -5,10 +5,12 @@ import be.ugent.idlab.knows.dataio.cores.StreamTestCore;
 import be.ugent.idlab.knows.dataio.streams.CSVSourceStream;
 import org.junit.Test;
 
+import java.nio.charset.StandardCharsets;
+
 public class CSVStreamTest extends StreamTestCore {
     @Test
     public void eval_0000() throws Exception {
-        Access access = makeLocalAccess("/csv/0000.csv", "src/test/resources", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/0000.csv", "src/test/resources", "csv", StandardCharsets.UTF_8);
         try (CSVSourceStream stream = new CSVSourceStream(access)) {
             runTest(stream, this::evaluate_0000);
         }
@@ -16,7 +18,7 @@ public class CSVStreamTest extends StreamTestCore {
 
     @Test
     public void eval_0001() throws Exception {
-        Access access = makeLocalAccess("/csv/0001.csv", "src/test/resources", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/0001.csv", "src/test/resources", "csv", StandardCharsets.UTF_8);
         try (CSVSourceStream stream = new CSVSourceStream(access)) {
             runTest(stream, this::evaluate_0001);
         }
@@ -24,7 +26,7 @@ public class CSVStreamTest extends StreamTestCore {
 
     @Test
     public void eval_1001_header_long() throws Exception {
-        Access access = makeLocalAccess("/csv/1001_header_long.csv", "src/test/resources", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/1001_header_long.csv", "src/test/resources", "csv", StandardCharsets.UTF_8);
         try (CSVSourceStream stream = new CSVSourceStream(access)) {
             runTest(stream, this::evaluate_1001_header_long);
         }
@@ -32,7 +34,7 @@ public class CSVStreamTest extends StreamTestCore {
 
     @Test
     public void eval_1001_header_short() throws Exception {
-        Access access = makeLocalAccess("/csv/1001_header_short.csv", "src/test/resources", "csv", "UTF-8");
+        Access access = makeLocalAccess("/csv/1001_header_short.csv", "src/test/resources", "csv", StandardCharsets.UTF_8);
         try (CSVSourceStream stream = new CSVSourceStream(access)) {
             runTest(stream, this::evaluate_1001_header_short);
         }