+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Use `Charset` instead of `String` for character encoding wherever possible.
+
+### Fixed
+- CSVWSourceIterator stopped after a certain number of bytes were parsed (GitLab [issue 21](https://gitlab.ilabt.imec.be/rml/proc/dataio/-/issues/21)). 
+  The bug originated from `CSVNullInjector`, it has completely been rewritten into `NewCSVNullInjector`.
+
 ## [1.0.4] - 2023-10-31
 
 ### Fixed